feat(Step 1): Add Initial Project Files based on VS Code C++ Docs
We start with a basic C++ project structure that does not use CMake.
This is based on the official VS Code C++ documentation found at:
https://code.visualstudio.com/docs/cpp/config-msvc

While we follow the docs linked above, we also make some modifications
like making the "msg" data structure more complex for upcoming
demonstration purposes when we add Python.

As mentioned, we do not use CMake for this tutorial. For a more complex
project, we would use CMake, following the instructions on VS Code's
documentation.