feat(Step 8): import Python module and use C++ code in Python -- Done!
We add the Pyhon/C++ module created in the previous step to our path,
and we are ready to import it and use C++ code in Python.

Our code seamlessly binds our python `list[dict[str, str]]` into C++
`std::vector<std::unordered_map<std::string, std::string>>` via
pybind11's automatic conversion described in the documentation:
https://pybind11.readthedocs.io/en/stable/advanced/cast/stl.html

However, as described in the documentation, this conversion copies
the data on each bind, so it may cause performance issues in some cases.
We need to be mindful about this. The documentation offers more
advanced techniques to avoid this.